Work around for the bug where animations got squished.
Corrected some spawning for the plant monsters.

Hitboxes are no longer affected by the animation pixel sizes. They are hardcoded to the entity for now as getting the width/height from a gif is too difficult with the current framework.

Entity#onGround() now checks only the one block below the entity rather than a left and right block.