Can now break cracked blocks with arrows.
Added option to enable the shader wave effects. Needs further optimisation but a good concept for now.

Added place holder death animations. Plays a timer on death that rotates the entity and shrinks them.

Optimisations to static decorations. The final texture is now saved to the decoration and rendered each time. Previously it would apply transformations to the image every draw call. Now it does this once when the decoration is added. Use updateTexture() for decorations if you do change their scale/rotation.

Added new sword and bow to test different damages based on the item the player is holding.