Minor update and bux fixes.
Entities now use attributes for their stats (attack damage, attack speed, movement speed ect). Attributes are stored as a hashmap with an assigned value. Only relevant attributes are store for the entity, for example a coin will not have an attack damage attribute.

As far as i can tell the collisions have been fully fixed. Turns out just increasing the quadtree collision boxes width and height by 1 pixel in each direction, made it so an object that is exactly on a division line between the boxes would be included in both of the adjacent boxes.